make sendTweet reusable
